A simple community-based website.
This is just a simple community based website me and my teammates made
for our semester project for the Web Engineering course. Everything is
made in HTML, CSS, JavaScript and PHP. The DBMS used is MySQL with
PHPMyAdmin.

The database is named 'thegamebook', and the SQL file to import it to
your own DBMS is in the root folder 'thegamebook.sql'.